+++ changelog.md
@@ -4,6 +4,7 @@ Liferay Plugin for IntelliJ
 All notable changes to this project will be documented in this file.
 
 ## 0.0.15
+- Feature: Ability to compare theme files with the original files from the parent theme
 
 ## 0.0.14
 - Feature: Resolve Resource Bundles